This is a combination of 23 commits. Basically this got the invitation working with payments properly, start invite and change date time using same UI, display request history properly, etc.
Switch the order. Go to askout activity directly after the invite button clicked.

Move the change date button out the post call back. What the heck was it doing there?

Switch to use the ChangeDateActivity

Change the ChangeDate activity from a dialog to a activity.

Fix up the ChangeDateActivity layout.
Remove the date_time_picker which is no longer used.

Add the location selector to the UI.

Reduce the margin.

Remove the AskThemOut activity and replace it with ChangeDateActivity.

Missed the deleting files in previous commit.

Add a NewDateActivity that is identical to ChangeDateActivity now.

A different title for the NewDateActivity

Add 2 handlers for the ok button click for new date and change date.

Make the NewDataActivity OK button point to the Amazon payment.

Payment activity return back the matchid created to the new date activity.

Add better handling for the result passed back from payment.
Call parent click OK handler for actually upload the date.

ChangeDateActivity get match id from previous activity

Log the date and time we got.

Add log of epoch

Post the epoch to server.

Change the request history UI.

Display time and location.

Add more info to the RequestHistoryActivity to display.
User the new server.

Missed a little on the URL.